+++ peernode/src/peer-node/consumer/mod.rs
@@ -20,7 +20,8 @@ pub async fn list_producers(file_hash: String, client: &mut MarketClient) -> Res
         if let Err(e) = writeln!(
             &mut producer_list,
             "Producer:\n  id: {}\n  Price: {}\n",
-            encoded_producer.as_str(), producer.price
+            encoded_producer.as_str(),
+            producer.price
         ) {
             eprintln!("Failed to write producer: {}", e);
             return Err(anyhow::anyhow!("Failed to write producer"));
