+++ peernode/src/peer-node/routes/bubble_guppies/settings.rs
@@ -1,8 +1,7 @@
 #![allow(non_snake_case)]
 use axum::{
     body::Body,
-    debug_handler,
-    extract::{Path, State},
+    extract::State,
     http::StatusCode,
     response::{IntoResponse, Response},
     routing::{get, put},
@@ -13,12 +12,16 @@ use serde::Deserialize;
 use crate::{store::Theme, ServerState};
 
 #[derive(Deserialize)]
+#[allow(dead_code)]
 struct SetSettings {
     theme: Theme,
     server: String,
 }
 
-async fn set_settings(State(state): State<ServerState>, Json(settings): Json<SetSettings>) -> Response {
+async fn set_settings(
+    State(state): State<ServerState>,
+    Json(settings): Json<SetSettings>,
+) -> Response {
     let mut config = state.config.lock().await;
     config.set_theme(settings.theme);
 
@@ -32,9 +35,11 @@ async fn get_settings(State(state): State<ServerState>) -> impl IntoResponse {
     let config = state.config.lock().await;
     Response::builder()
         .status(StatusCode::OK)
-        .body(Body::from(format!(r#"
+        .body(Body::from(format!(
+            r#"
 {{"theme": "{:?}","server":"rust"}}
-"#, config.get_theme()
+"#,
+            config.get_theme()
         )))
         .unwrap()
 }
