+++ peernode/src/peer-node/routes/manta/mod.rs
@@ -2,8 +2,8 @@ pub mod mining;
 pub mod stats;
 pub mod wallet;
 
-use axum::Router;
 use crate::ServerState;
+use axum::Router;
 
 pub fn routes() -> Router<ServerState> {
     Router::new()
